Move settings module to sync
