--- settings/settings-api/build.gradle
@@ -1,15 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
-}
-
-apply from: "$rootProject.projectDir/gradle/android-library.gradle"
-
-dependencies {
-    implementation Kotlin.stdlib.jdk7
-    implementation AndroidX.core.ktx
-}
-
-android {
-    namespace 'com.duckduckgo.settings.api'
-}
