--- settings/settings-impl/build.gradle
@@ -1,90 +1,0 @@
-/*
- * Copyright (c) 2021 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
-    id 'com.squareup.anvil'
-    id 'kotlin-kapt'
-}
-
-apply from: "$rootProject.projectDir/gradle/android-library.gradle"
-
-dependencies {
-    implementation project(path: ':di')
-    implementation project(path: ':common-utils')
-    implementation project(path: ':common-ui')
-    implementation project(path: ':settings-api')
-    implementation project(path: ':sync-api')
-
-    anvil project(path: ':anvil-compiler')
-    implementation project(path: ':anvil-annotations')
-
-    implementation AndroidX.appCompat
-    implementation Google.android.material
-    implementation JakeWharton.timber
-
-    implementation KotlinX.coroutines.core
-
-    implementation Square.retrofit2.converter.moshi
-    implementation "com.squareup.moshi:moshi-kotlin:_"
-    implementation Google.dagger
-    implementation AndroidX.core.ktx
-
-    // Room
-    implementation AndroidX.room.runtime
-    implementation AndroidX.room.ktx
-    kapt AndroidX.room.compiler
-
-    implementation "com.jakewharton.threetenabp:threetenabp:_"
-    testImplementation "org.threeten:threetenbp:_"
-
-    // Testing dependencies
-    testImplementation "org.mockito.kotlin:mockito-kotlin:_"
-    testImplementation Testing.junit4
-    testImplementation AndroidX.core
-    testImplementation AndroidX.test.ext.junit
-    testImplementation "androidx.test:runner:_"
-    testImplementation Testing.robolectric
-    testImplementation CashApp.turbine
-    testImplementation AndroidX.archCore.testing
-    testImplementation "org.jetbrains.kotlin:kotlin-reflect:_"
-    testImplementation 'org.json:json:_'
-
-    testImplementation project(path: ':common-test')
-
-    testImplementation (KotlinX.coroutines.test) {
-        // https://github.com/Kotlin/kotlinx.coroutines/issues/2023
-        // conflicts with mockito due to direct inclusion of byte buddy
-        exclude group: "org.jetbrains.kotlinx", module: "kotlinx-coroutines-debug"
-    }
-}
-
-android {
-    anvil {
-        generateDaggerFactories = true // default is false
-    }
-    lint {
-        baseline file("lint-baseline.xml")
-    }
-    testOptions {
-        unitTests {
-            includeAndroidResources = true
-        }
-    }
-    namespace 'com.duckduckgo.settings.impl'
-}
-
