--- settings/settings-impl/src/main/java/com/duckduckgo/settings/impl/SettingsSyncMetadataDao.kt
@@ -1,46 +1,0 @@
-/*
- * Copyright (c) 2023 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.duckduckgo.settings.impl
-
-import androidx.room.*
-import kotlinx.coroutines.flow.*
-
-@Dao
-interface SettingsSyncMetadataDao {
-
-    @Query("select * from settings_sync_meta where `key` = :key")
-    fun get(key: String): SettingsSyncMetadataEntity?
-
-    @Query("select * from settings_sync_meta")
-    fun getAllObservable(): Flow<List<SettingsSyncMetadataEntity>>
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    fun addOrUpdate(settingsSyncMetadataEntity: SettingsSyncMetadataEntity): Long
-
-    @Query("select * from settings_sync_meta where modified_at > :since")
-    fun getChangesSince(since: String): List<SettingsSyncMetadataEntity>
-
-    @Query("Delete from settings_sync_meta")
-    fun removeAll()
-
-    @Transaction
-    fun initialize(keys: List<String>) {
-        removeAll()
-        keys.map { SettingsSyncMetadataEntity(key = it, deleted_at = null, modified_at = SyncDateProvider.now()) }
-            .forEach { addOrUpdate(it) }
-    }
-}
