--- settings/settings-impl/src/main/java/com/duckduckgo/settings/impl/SettingsSyncMetadataEntity.kt
@@ -1,28 +1,0 @@
-/*
- * Copyright (c) 2023 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.duckduckgo.settings.impl
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-import java.util.*
-
-@Entity(tableName = "settings_sync_meta")
-data class SettingsSyncMetadataEntity(
-    @PrimaryKey val key: String,
-    var modified_at: String?, // should follow iso8601 format
-    var deleted_at: String?, // should follow iso8601 format
-)
