--- settings/settings-impl/src/main/java/com/duckduckgo/settings/impl/StoreModule.kt
@@ -1,45 +1,0 @@
-/*
- * Copyright (c) 2023 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.duckduckgo.settings.impl
-
-import android.content.Context
-import androidx.room.Room
-import com.duckduckgo.di.scopes.AppScope
-import com.squareup.anvil.annotations.ContributesTo
-import dagger.Module
-import dagger.Provides
-import dagger.SingleInstanceIn
-
-@Module
-@ContributesTo(AppScope::class)
-class StoreModule {
-
-    @SingleInstanceIn(AppScope::class)
-    @Provides
-    fun provideSettingsDatabase(context: Context): SettingsDatabase {
-        return Room.databaseBuilder(context, SettingsDatabase::class.java, "settings.db")
-            .fallbackToDestructiveMigration()
-            .addMigrations(*ALL_MIGRATIONS)
-            .build()
-    }
-
-    @SingleInstanceIn(AppScope::class)
-    @Provides
-    fun provideSettingsSyncMetadataDao(database: SettingsDatabase): SettingsSyncMetadataDao {
-        return database.settingsSyncDao()
-    }
-}
