--- settings/settings-impl/src/test/java/com/duckduckgo/settings/impl/SyncableSettingsPluginPoint.kt
@@ -1,28 +1,0 @@
-/*
- * Copyright (c) 2023 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.duckduckgo.settings.impl
-
-import com.duckduckgo.app.global.plugins.*
-import com.duckduckgo.settings.api.*
-
-class SyncableSettingsPluginPoint(
-    val syncableSettings: MutableList<SyncableSetting>,
-) : PluginPoint<SyncableSetting> {
-    override fun getPlugins(): Collection<SyncableSetting> {
-        return syncableSettings
-    }
-}
