+++ sync/sync-settings-api/src/main/java/com/duckduckgo/sync/settings/api/SyncSettingsListener.kt
@@ -1,0 +1,30 @@
+/*
+ * Copyright (c) 2023 DuckDuckGo
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.duckduckgo.sync.settings.api
+
+/**
+ * Interface to notify when a setting was changed by the user.
+ */
+interface SyncSettingsListener {
+
+    /**
+     * Should be called when a setting was changed by the user.
+     * @param settingKey the unique identifier of the setting that was changed.
+     * It should be the same as @see SyncableSetting.key
+     */
+    fun onSettingChanged(settingKey: String)
+}
