+++ sync/sync-settings-impl/src/test/resources/json/settings_with_deleted_object_response.json
