+++ sync/sync-settings-impl/src/test/resources/json/settings_with_values_response.json
